Se agrega monedero electronico con base al numero de celular del cliente

+++ src/main/java/com/school/app/dto/requets/SaleItemRequest.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 
 public record SaleItemRequest(
         UUID productId,
-        Integer quantity,
+        BigDecimal quantity,
         BigDecimal unitPrice
 ) {
 }
